camera/test_camera.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for mozCameras</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body onload="doTest()">

<script class="testbody" type="text/javascript;version=1.7">

function doTest() {
  //navigator.mozCameras.getCamera(function() {alert('onsucces')}, function() {alert('onerror')});;
  //ok(true, "navigator.mozCameras.getCamera should not throw");

try {
//  cameras = navigator.mozCameras.getListOfCameras();
//  ok(cameras[0] == "back", "the back camera is available");
//  ok(true, "navigator.mozCameras.getListOfCameras should not throw");
//  cameras[0].viewfinder.pause();
//  cameras[0].viewfinder.mozSrcObject = null;

} catch(e) {
//  ok(false, "navigator.mozCameras.getListOfCameras should not throw");
}
SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();

</script>
</body>

</html>
